Move processing class to another file. PEP8 fixes.
